Improved duration formatting
